docs(Layout): move media query docs to layout component docs
