+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout.mdx
@@ -12,6 +12,8 @@ tabs:
 breadcrumb:
   - text: Layout
     href: /uilib/components/layout/
+redirect_from:
+  - /uilib/usage/layout
 ---
 
 import LayoutInfo from 'Docs/uilib/components/layout/info'
