+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout/Examples.tsx
@@ -1,5 +1,8 @@
+import React from 'react'
+import styled from '@emotion/styled'
 import ComponentBox from '../../../../shared/tags/ComponentBox'
-import { Layout, Slider } from '@dnb/eufemia/src'
+import MediaQuery from '@dnb/eufemia/src/shared/MediaQuery'
+import { Layout, Slider, Code, Button } from '@dnb/eufemia/src'
 import {
   TestElement,
   Field,
@@ -8,7 +11,7 @@ import {
 } from '@dnb/eufemia/src/extensions/forms'
 import { defaultBreakpoints } from '@dnb/eufemia/src/shared/MediaQueryUtils'
 import { defaultQueries } from '@dnb/eufemia/src/shared/useMedia'
-import styled from '@emotion/styled'
+import { useMedia, useMediaQuery } from '@dnb/eufemia/src/shared'
 
 export const LayoutComponents = () => {
   return (
@@ -186,3 +189,95 @@ export const HorizontalAutoSize = () => {
     </ComponentBox>
   )
 }
+
+const useWindowWidth = () => {
+  const [innerWidth, setWidth] = React.useState(
+    typeof window !== 'undefined' ? window.innerWidth : 0,
+  )
+
+  React.useEffect(() => {
+    const resizeHandler = () => {
+      setWidth(window.innerWidth)
+    }
+    window.addEventListener('resize', resizeHandler)
+    return () => window.removeEventListener('resize', resizeHandler)
+  }, [])
+
+  return { innerWidth }
+}
+
+export const MediaQueryUseMedia = () => (
+  <ComponentBox scope={{ useMedia, useWindowWidth }} hideCode>
+    {() => {
+      const Playground = () => {
+        const { isSmall, isMedium, isLarge, isSSR } = useMedia()
+        const { innerWidth } = useWindowWidth()
+
+        return (
+          <Code>
+            <pre>
+              {JSON.stringify(
+                { isSmall, isMedium, isLarge, isSSR, innerWidth },
+                null,
+                2,
+              )}
+            </pre>
+          </Code>
+        )
+      }
+      return <Playground />
+    }}
+  </ComponentBox>
+)
+
+export const MediaQueryLiveExample = () => (
+  <ComponentBox scope={{ MediaQuery, useMediaQuery }} hideCode>
+    {() => {
+      const Playground = () => {
+        const [query, updateQuery] = React.useState({
+          screen: true,
+          not: true,
+          min: 'small',
+          max: 'large',
+        })
+
+        const match1 = useMediaQuery({
+          matchOnSSR: true,
+          when: query,
+        })
+        const match2 = useMediaQuery({
+          matchOnSSR: true,
+          not: true,
+          when: query,
+        })
+
+        React.useEffect(() => {
+          console.log('mediaQuery:', match1, match2)
+        }, [match1, match2])
+
+        return (
+          <>
+            <Button
+              onClick={() => {
+                updateQuery({
+                  ...query,
+                  screen: !query.screen,
+                })
+              }}
+              right
+            >
+              Switch
+            </Button>
+            <MediaQuery when={query}>
+              <Code>when</Code>
+            </MediaQuery>
+            <MediaQuery not when={query}>
+              <Code>not when</Code>
+            </MediaQuery>
+          </>
+        )
+      }
+      return <Playground />
+    }}
+  </ComponentBox>
+)
