+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout/info.mdx
@@ -38,3 +38,32 @@ render(
 - **[MainHeading](/uilib/extensions/forms/extended-features/Form/MainHeading)** is a standardized main heading for sections, ensuring default layout, spacing etc.
 
 - **[SubHeading](/uilib/extensions/forms/extended-features/Form/SubHeading)** is a standardized sub heading for sections, ensuring default layout, spacing etc.
+
+You may use the Eufemia [Layout](/uilib/components/layout/) component for your application layouts and [form](/uilib/extensions/forms)-views.
+
+Or you may just use **CSS Flexbox** or **CSS Grid**. But please, do not use third-party layout libraries.
+
+It is your responsibility to ensure 100% consistency and compatibility.
+
+## Grid
+
+UX designers are using a 12 column system, along with a 4 and 6 column system, during their design processes.
+
+There are plans to implement support for this in the future.
+
+## Responsiveness
+
+In short, only use `rem` for layouts and spacing and make sure ...
+
+- you always use the nearest half `rem` value, like _0.5rem_, _1rem_ or _1.5rem_ and so forth.
+- you always get a **total computed height** within the grid.
+
+This results in maintaining the integrity of the **8px** base grid.
+
+### Smaller Units
+
+Sometimes you may need a compensation of only a few pixels. Heres how to calculate the correct _rem_ values:
+
+- 1px = `1/16x1` = **0.0625rem**
+- 2px = `1/16x2` = **0.125rem**
+- And so on ...
