+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout/media-queries.mdx
@@ -1,12 +1,11 @@
 ---
 title: 'Media Queries'
 order: 2
+redirect_from:
+  - /uilib/usage/layout/media-queries
 ---
 
-import {
-  MediaQueryLiveExample,
-  MediaQueryUseMedia,
-} from 'dnb-design-system-portal/src/docs/uilib/usage/layout/Examples'
+import { MediaQueryLiveExample, MediaQueryUseMedia } from './Examples'
 
 # Media Queries and Breakpoints
