+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout/spacing-table.mdx
