+++ packages/dnb-design-system-portal/src/docs/uilib/components/layout/spacing.mdx
@@ -1,9 +1,11 @@
 ---
 title: 'Spacing'
 order: 1
+redirect_from:
+  - /uilib/usage/layout/spacing
 ---
 
-import SpacingTable from 'Docs/uilib/usage/layout/spacing-table.mdx'
+import SpacingTable from 'Docs/uilib/components/layout/spacing-table.mdx'
 
 # Spacing
