+++ packages/dnb-design-system-portal/src/docs/uilib/components/space/info.mdx
@@ -2,7 +2,7 @@
 showTabs: true
 ---
 
-import SpacingTable from 'Docs/uilib/usage/layout/spacing-table.mdx'
+import SpacingTable from 'Docs/uilib/components/layout/spacing-table.mdx'
 import {
   SpaceExamplesSameResult1,
   SpaceExamplesSameResult2,
